Move Pure/Impure RenderLoop logic
