+++ core/js/src/main/scala/eu/joaocosta/minart/backend/JsRenderLoop.scala
@@ -8,7 +8,7 @@ import scala.scalajs.js.timers
 
 import eu.joaocosta.minart.core._
 
-object JsRenderLoop extends RenderLoop.ImpureRenderLoop {
+object JsRenderLoop extends ImpureRenderLoop {
   def finiteRenderLoop[S](
     canvasManager: CanvasManager,
     initialState: S,
