+++ core/js/src/test/scala/eu/joaocosta/minart/backend/JsRenderLoopSpec.scala
@@ -1,9 +1,10 @@
 package eu.joaocosta.minart.backend
 
+import eu.joaocosta.minart.backend._
 import eu.joaocosta.minart.core._
 
 class JsRenderLoopSpec extends RenderLoopTests {
-  lazy val renderLoop: RenderLoop.ImpureRenderLoop = JsRenderLoop
+  lazy val renderLoop: ImpureRenderLoop = JsRenderLoop
   lazy val renderLoopName: String = "JsRenderLoop"
   override lazy val testLoop: Boolean = false
 }
