+++ core/jvm/src/test/scala/eu/joaocosta/minart/backend/JavaRenderLoopSpec.scala
@@ -1,8 +1,9 @@
 package eu.joaocosta.minart.backend
 
+import eu.joaocosta.minart.backend._
 import eu.joaocosta.minart.core._
 
 class JavaRenderLoopSpec extends RenderLoopTests {
-  lazy val renderLoop: RenderLoop.ImpureRenderLoop = JavaRenderLoop
+  lazy val renderLoop: ImpureRenderLoop = JavaRenderLoop
   lazy val renderLoopName: String = "JavaRenderLoop"
 }
