+++ core/shared/src/main/scala/eu/joaocosta/minart/core/RenderLoop.scala
@@ -66,26 +66,9 @@ trait RenderLoop[F1[-_, +_], F2[-_, -_, +_]] {
 }
 
 object RenderLoop {
-
-  trait ImpureRenderLoop extends RenderLoop[Function1, Function2] {
-    def infiniteRenderLoop[S](
-      canvasManager: CanvasManager,
-      initialState: S,
-      renderFrame: (Canvas, S) => S,
-      frameRate: FrameRate): Unit =
-      finiteRenderLoop(canvasManager, initialState, renderFrame, (_: S) => false, frameRate)
-
-    def infiniteRenderLoop(
-      canvasManager: CanvasManager,
-      renderFrame: Canvas => Unit,
-      frameRate: FrameRate): Unit =
-      infiniteRenderLoop(canvasManager, (), (c: Canvas, _: Unit) => renderFrame(c), frameRate)
-
-  }
-
   /**
-   * Returns [[RenderLoop]] for the default backend for the target platform.
+   * Returns a [[RenderLoop]] for the default backend for the target platform.
    */
-  def default()(implicit d: DefaultBackend[Any, ImpureRenderLoop]): ImpureRenderLoop =
+  def default()(implicit d: DefaultBackend[Any, RenderLoop[Function1, Function2]]): RenderLoop[Function1, Function2] =
     d.defaultValue(())
 }
