+++ examples/purecolorsquare/shared/src/main/scala/eu/joaocosta/minart/examples/PureColorSquare.scala
@@ -2,6 +2,7 @@ package eu.joaocosta.minart.examples
 
 import eu.joaocosta.minart.core._
 import eu.joaocosta.minart.pure._
+import eu.joaocosta.minart.pure.backend._
 
 object PureColorSquare extends MinartApp {
   type State = Unit
