+++ pure/shared/src/main/scala/eu/joaocosta/minart/pure/MinartApp.scala
@@ -1,6 +1,7 @@
 package eu.joaocosta.minart.pure
 
 import eu.joaocosta.minart.core._
+import eu.joaocosta.minart.pure.backend._
 
 /** Entrypoint for pure Minart applications. */
 trait MinartApp {
